+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Improves element string representation
 - Adds `execution.output_filepath` and `execution.component_directory`
+- Reject stale context matches in `index.find`
 
 # v4.6.2